refactor(agentic-ai): fold the Entra ID Foundry scope resolver into FoundryCredentialResolver
EntraIdFoundryScopeResolver was extracted to be shared between the
Anthropic and OpenAI Foundry credential resolvers, but those have since
been unified into a single FoundryCredentialResolver, leaving it with
one consumer in its own package and a javadoc claiming otherwise.

Collapses scopeFor/resolveScope into a single private
scopeFor(authorityHost, scopeOverride) so callers no longer compose the
two, and makes both scope constants private again.
